+++ src/games/application/games.service.ts
@@ -1,12 +1,8 @@
-import { BadRequestException, Inject, Injectable, NotFoundException } from '@nestjs/common';
+import { Inject, Injectable } from '@nestjs/common';
 
-import { ClientAnswer } from '../domain/game-client-answers.interface';
 import { GameOptions } from '../domain/game-options.entity';
-import { GameProblem } from '../domain/game-problem.entity';
-import { calculateServerScore, ProblemDifficulty } from '../domain/game.business-rules';
 import { GAME_REPOSITORY, IGameRepository } from '../domain/games.repository.interface';
 import { UserMistakeAnalysis } from '../domain/user-mistake-analysis.entity';
-import { CreateGameSessionServiceRequestDto } from './service-dto/create-game-session.service-dto';
 
 @Injectable()
 export class GamesService {
@@ -21,112 +17,6 @@ export class GamesService {
     return GameOptions.from(categories);
   }
 
-  /**
-   *
-   * @description 게임세션 저장
-   */
-  async createGameSession(command: CreateGameSessionServiceRequestDto): Promise<bigint> {
-    const serverScore = await this.validateAndCalculateScore(
-      command.categoryId,
-      command.clientAnswers,
-    );
-
-    return this.gameRepository.saveGameSession({
-      categoryId: command.categoryId,
-      difficultyMode: command.difficultyMode,
-      score: serverScore,
-      totalProblemCount: command.clientAnswers.length,
-      correctProblemCount: command.clientAnswers.filter((a) => a.solved).length,
-      logs: command.clientAnswers.map((a) => ({
-        problemId: BigInt(a.problemId),
-        inputs: a.inputs,
-        isSolved: a.solved,
-        tryCount: a.inputs.length,
-      })),
-    });
-  }
-
-  /**
-   * @description 클라이언트 게임 데이터 검증 후 서버 점수 계산
-   */
-  private async validateAndCalculateScore(
-    categoryId: number,
-    clientAnswers: ClientAnswer[],
-  ): Promise<number> {
-    this.validateAnswerIntegrity(clientAnswers);
-
-    const [, problems] = await Promise.all([
-      this.validateCategory(categoryId),
-      this.validateAndMatchedGameProblems(clientAnswers),
-    ]);
-    return this.calculateScore(clientAnswers, problems);
-  }
-
-  /**
-   *
-   * @description 게임 카테고리 검증
-   */
-  private async validateCategory(categoryId: number): Promise<void> {
-    const exists = await this.gameRepository.categoryExists(categoryId);
-    if (!exists) {
-      throw new NotFoundException('존재하지 않는 카테고리입니다.');
-    }
-  }
-
-  /**
-   * @description 클라이언트가 푼 게임문제 검증
-   */
-  private async validateAndMatchedGameProblems(
-    clientAnswers: ClientAnswer[],
-  ): Promise<GameProblem[]> {
-    const problemIds = clientAnswers.map((a) => BigInt(a.problemId));
-    const uniqueProblemIds = [...new Set(problemIds)];
-
-    if (uniqueProblemIds.length !== problemIds.length) {
-      throw new BadRequestException('clientAnswers에 중복된 problemId가 포함되어 있습니다.');
-    }
-
-    const problems = await this.gameRepository.findProblemsByIds(uniqueProblemIds);
-
-    if (problems.length !== uniqueProblemIds.length) {
-      const foundIds = new Set(problems.map((p) => p.id));
-      const missingIds = uniqueProblemIds.filter((id) => !foundIds.has(id));
-      throw new NotFoundException(
-        `존재하지 않는 문제 ID가 포함되어 있습니다. (problemId: ${missingIds.join(', ')})`,
-      );
-    }
-
-    return problems;
-  }
-
-  /**
-   * @description 클라이언트 문제풀이 입력데이터가 정답인지 검증
-   */
-  private validateAnswerIntegrity(clientAnswers: ClientAnswer[]): void {
-    const invalid = clientAnswers.find(
-      (a) => a.solved && !a.inputs.some((input) => input.isCorrect),
-    );
-    if (invalid) {
-      throw new BadRequestException(
-        '데이터 무결성 오류: solved가 true이지만 정답 처리된 입력이 없습니다.',
-      );
-    }
-  }
-
-  /**
-   * @description 클라이언트 점수 데이터가 올바르게 계산됐는지 재검증
-   */
-  private calculateScore(clientAnswers: ClientAnswer[], problems: GameProblem[]): number {
-    const problemMap = new Map(problems.map((p) => [p.id, p]));
-    const solvedDifficulties: ProblemDifficulty[] = clientAnswers
-      .filter((a) => a.solved)
-      .map((a) => {
-        const problem = problemMap.get(BigInt(a.problemId))!;
-        return problem.difficulty;
-      });
-
-    return calculateServerScore(solvedDifficulties);
-  }
   /*
    * @description 사용자의 명령어, 카테고리 실수 분석 조회
    * - 자주 틀린 명령어 Top 5 (서브 카테고리 별)
